+++ .github/workflows/releases-comparator.yml
@@ -91,6 +91,8 @@ jobs:
             "com/hedera/hapi/node/state/schedule/Schedule.java"
             "com/swirlds/state/spi/ReadableKVStateBase.java"
             "com/swirlds/state/spi/WritableKVStateBase.java"
+            "com/hedera/node/app/service/contract/impl/exec/processors/CustomMessageCallProcessor.java"
+            "com/hedera/node/app/service/contract/impl/state/DispatchingEvmFrameState.java"
           )
 
           # Temp directories to store cloned repos
