+++ web3/src/main/java/org/hiero/mirror/web3/evm/contracts/execution/traceability/AbstractOpcodeTracer.java
@@ -2,34 +2,21 @@
 
 package org.hiero.mirror.web3.evm.contracts.execution.traceability;
 
-import com.hedera.node.app.service.contract.impl.exec.systemcontracts.HederaSystemContract;
 import com.hederahashgraph.api.proto.java.ResponseCodeEnum;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import lombok.Getter;
-import lombok.Setter;
 import org.apache.tuweni.bytes.Bytes;
 import org.hiero.mirror.common.domain.contract.ContractAction;
 import org.hiero.mirror.web3.common.ContractCallContext;
 import org.hiero.mirror.web3.convert.BytesDecoder;
-import org.hyperledger.besu.datatypes.Address;
 import org.hyperledger.besu.evm.frame.MessageFrame;
-import org.hyperledger.besu.evm.precompile.PrecompiledContract;
 import org.springframework.util.CollectionUtils;
 
 public abstract class AbstractOpcodeTracer {
 
-    @Getter
-    @Setter
-    protected Map<Address, HederaSystemContract> systemContracts = new HashMap<>();
-
-    protected Map<Address, PrecompiledContract> hederaPrecompiles = new HashMap<>();
-
-    protected List<Bytes> captureMemory(final MessageFrame frame, final OpcodeTracerOptions options) {
+    protected final List<Bytes> captureMemory(final MessageFrame frame, final OpcodeTracerOptions options) {
         if (!options.isMemory()) {
             return Collections.emptyList();
         }
@@ -43,7 +30,7 @@ public abstract class AbstractOpcodeTracer {
         return memory;
     }
 
-    protected List<Bytes> captureStack(final MessageFrame frame, final OpcodeTracerOptions options) {
+    protected final List<Bytes> captureStack(final MessageFrame frame, final OpcodeTracerOptions options) {
         if (!options.isStack()) {
             return Collections.emptyList();
         }
@@ -57,7 +44,7 @@ public abstract class AbstractOpcodeTracer {
         return stack;
     }
 
-    protected Optional<Bytes> getRevertReasonFromContractActions(final ContractCallContext context) {
+    protected final Optional<Bytes> getRevertReasonFromContractActions(final ContractCallContext context) {
         final var contractActions = context.getContractActions();
 
         if (CollectionUtils.isEmpty(contractActions)) {
@@ -69,18 +56,6 @@ public abstract class AbstractOpcodeTracer {
                 .map(action -> Bytes.of(action.getResultData()))
                 .map(this::formatRevertReason)
                 .findFirst();
-    }
-
-    protected boolean isCallToHederaPrecompile(
-            final MessageFrame frame, final Map<Address, PrecompiledContract> hederaPrecompiles) {
-        final var recipientAddress = frame.getRecipientAddress();
-        return hederaPrecompiles.containsKey(recipientAddress);
-    }
-
-    protected boolean isCallToSystemContracts(
-            final MessageFrame frame, final Map<Address, HederaSystemContract> systemContracts) {
-        final var recipientAddress = frame.getRecipientAddress();
-        return systemContracts.containsKey(recipientAddress);
     }
 
     /**
