+++ web3/src/main/java/org/hiero/mirror/web3/evm/contracts/execution/traceability/OpcodeActionTracer.java
@@ -5,6 +5,7 @@ package org.hiero.mirror.web3.evm.contracts.execution.traceability;
 import com.hedera.hapi.streams.ContractActionType;
 import com.hedera.hapi.streams.ContractActions;
 import com.hedera.node.app.service.contract.impl.exec.ActionSidecarContentTracer;
+import com.hedera.node.app.service.contract.impl.exec.systemcontracts.HederaSystemContract;
 import com.hedera.node.app.service.contract.impl.state.ProxyWorldUpdater;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
@@ -12,11 +13,14 @@ import jakarta.inject.Named;
 import java.util.Collections;
 import java.util.Map;
 import java.util.TreeMap;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
 import lombok.CustomLog;
+import lombok.Getter;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.tuweni.bytes.Bytes;
 import org.hiero.mirror.web3.common.ContractCallContext;
+import org.hyperledger.besu.datatypes.Address;
 import org.hyperledger.besu.evm.ModificationNotAllowedException;
 import org.hyperledger.besu.evm.frame.MessageFrame;
 import org.hyperledger.besu.evm.operation.Operation.OperationResult;
@@ -24,6 +28,9 @@ import org.hyperledger.besu.evm.operation.Operation.OperationResult;
 @Named
 @CustomLog
 public class OpcodeActionTracer extends AbstractOpcodeTracer implements ActionSidecarContentTracer {
+
+    @Getter
+    private final Map<Address, HederaSystemContract> systemContracts = new ConcurrentHashMap<>();
 
     @Override
     public void tracePostExecution(@NonNull final MessageFrame frame, @NonNull final OperationResult operationResult) {
@@ -113,5 +120,17 @@ public class OpcodeActionTracer extends AbstractOpcodeTracer implements ActionSi
     @Override
     public ContractActions contractActions() {
         return null;
+    }
+
+    public void setSystemContracts(final Map<Address, HederaSystemContract> systemContracts) {
+        if (systemContracts != null) {
+            this.systemContracts.putAll(systemContracts);
+        }
+    }
+
+    private boolean isCallToSystemContracts(
+            final MessageFrame frame, final Map<Address, HederaSystemContract> systemContracts) {
+        final var recipientAddress = frame.getRecipientAddress();
+        return systemContracts.containsKey(recipientAddress);
     }
 }
